requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
discord-pretty-help==2.0.7
discord.py==2.4.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
pycparser==2.22
PyNaCl==1.5.0
redis==5.0.7
requests==2.32.3
sniffio==1.3.1
spotipy==2.24.0
urllib3==2.2.2
validators==0.33.0
yarl==1.9.4
youtube-search-python==1.6.6
